+++ src/common/core/pro/jobs/download-pro-plugins.py
@@ -95,7 +95,6 @@ def install_plugin(plugin_path: Path, db, preview: bool = True) -> bool:
 try:
     db = Database(LOGGER, sqlalchemy_string=getenv("DATABASE_URI"))
     db_metadata = db.get_metadata()
-    db_config = db.get_config()
     current_date = datetime.now()
     pro_license_key = getenv("PRO_LICENSE_KEY", "").strip()
 
@@ -110,6 +109,7 @@ try:
     headers = {"User-Agent": f"BunkerWeb/{data['version']}"}
     default_metadata = {
         "is_pro": False,
+        "pro_license": pro_license_key,
         "pro_expire": None,
         "pro_status": "invalid",
         "pro_overlapped": False,
@@ -157,7 +157,7 @@ try:
 
     # ? If we already checked today, skip the check and if the metadata is the same, skip the check
     if (
-        pro_license_key == db_config["PRO_LICENSE_KEY"]
+        pro_license_key == db_metadata.get("pro_license", "")
         and metadata.get("is_pro", False) == db_metadata["is_pro"]
         and db_metadata["last_pro_check"]
         and current_date.replace(hour=0, minute=0, second=0, microsecond=0) == db_metadata["last_pro_check"].replace(hour=0, minute=0, second=0, microsecond=0)